fixed win()
To compensate for the random placement of bombs, and the chance that a bomb might not be placed or replaced, I've added a BOMBS, and a counter in the place_bombs() to keep track of the actual bombs placed.  To simplify the win()  we only need to know if they cleared all Non-bomb squares.